+++ src/javascript/components/panel/panel.js
@@ -1,5 +1,4 @@
 import {puffin} from '@mkenzo_8/puffin'
-import RunningConfig from 'RunningConfig'
 
 const PanelBody  = puffin.element(`
     <div click="$focusPanel" class="${