+++ src/javascript/components/window/side.menu.js
@@ -34,7 +34,7 @@ const SideMenu = puffin.element(`
             display:block;
             padding:20px;
             overflow:auto;
-            box-shadow:0px 3px 10px rgba(0,0,0,0.2);
+            box-shadow:0px 0px 10px rgba(0,0,0,0.2);
         }
         & > div:nth-child(1) > h1 {
             overflow:hidden;