+++ src/javascript/constructors/contextmenu.js
@@ -1,8 +1,6 @@
 import { puffin } from '@mkenzo_8/puffin'
 import ThemeProvider from 'ThemeProvider'
 
-
-
 function ContextMenu({
     list,
     parent,
@@ -18,6 +16,7 @@ function ContextMenu({
             color:white;
             border-radius:5px;
             box-shadow:0px 0px 3px rgba(0,0,0,0.2);
+            display:block;
         }
         & > button{
             background:{{contextmenuButtonBackground}};
@@ -26,10 +25,21 @@ function ContextMenu({
             padding:6px;
             outline:0;
             border-radius:5px;
+            display:block;
+            width:100%;
+            text-align:left;
         }
         & >  button:hover{
             background:{{contextmenuButtonHoveringBackground}};
             color:{{contextmenuButtonHoveringText}};
+        }
+        & >  span{
+            height:1.5px;
+            border-radius:25px;
+            width:95%;
+            display:block;
+            background:{{contextmenuDivider}};
+            margin:3px auto;
         }
 
     `
@@ -44,7 +54,12 @@ function ContextMenu({
                 ${(function(){
                     let content = "";
                         list.map((option,index)=>{
-                            content += `<button click="$${index}">${option.label}</button>`
+                            if(option.label != undefined){
+                                content += `<button click="$${index}">${option.label}</button>`
+                            }else{
+                                content += `<span></span>`
+                            }
+                            
                         })
                     return content
                 })()}